Move common router components to separate file
