+++ sdk/tests/router/config/test_router_ensembler_config.py
@@ -1,5 +1,6 @@
 import turing.generated.models
 import turing.router.config.router_ensembler_config
+import turing.router.config.common
 import turing.router.config.resource_request
 import pytest
 
@@ -76,7 +77,7 @@ def test_create_router_ensembler_config(id, type, standard_config, docker_config
             "500ms",
             5120,
             [
-                turing.router.config.router_ensembler_config.EnvVar(
+                turing.router.config.common.EnvVar(
                     name="env_name",
                     value="env_val")
             ],
@@ -124,7 +125,7 @@ def test_create_docker_router_ensembler_config(
             "500ms",
             5120,
             [
-                turing.router.config.router_ensembler_config.EnvVar(
+                turing.router.config.common.EnvVar(
                     name="env_name",
                     value="env_val")
             ],
@@ -172,7 +173,7 @@ def test_create_docker_router_ensembler_config_with_invalid_image(
             "500ms",
             5120,
             [
-                turing.router.config.router_ensembler_config.EnvVar(
+                turing.router.config.common.EnvVar(
                     name="env_name",
                     value="env_val")
             ],
@@ -222,7 +223,7 @@ def test_set_docker_router_ensembler_config_with_invalid_image(
             "500ms",
             5120,
             [
-                turing.router.config.router_ensembler_config.EnvVar(
+                turing.router.config.common.EnvVar(
                     name="env_name",
                     value="env_val")
             ],
@@ -272,7 +273,7 @@ def test_set_docker_router_ensembler_config_with_valid_image(
             "500ks",
             5120,
             [
-                turing.router.config.router_ensembler_config.EnvVar(
+                turing.router.config.common.EnvVar(
                     name="env_name",
                     value="env_val")
             ],
@@ -320,7 +321,7 @@ def test_create_docker_router_ensembler_config_with_invalid_timeout(
             "500ms",
             5120,
             [
-                turing.router.config.router_ensembler_config.EnvVar(
+                turing.router.config.common.EnvVar(
                     name="env_name",
                     value="env_val")
             ],
@@ -370,7 +371,7 @@ def test_set_docker_router_ensembler_config_with_invalid_timeout(
             "1000ms",
             5120,
             [
-                turing.router.config.router_ensembler_config.EnvVar(
+                turing.router.config.common.EnvVar(
                     name="env_name",
                     value="env_val")
             ],
@@ -420,12 +421,12 @@ def test_set_docker_router_ensembler_config_with_valid_timeout(
             "500ms",
             5120,
             [
-                turing.router.config.router_ensembler_config.EnvVar(
+                turing.router.config.common.EnvVar(
                     name="#@!#!",
                     value="env_val")
             ],
             "secret-name-for-google-service-account",
-            turing.router.config.router_ensembler_config.InvalidEnvironmentVariableNameException
+            turing.router.config.common.InvalidEnvironmentVariableNameException
         )
     ])
 def test_create_docker_router_ensembler_config_with_invalid_env(
@@ -456,7 +457,7 @@ def test_create_docker_router_ensembler_config_with_invalid_env(
     "new_env,id,image,resource_request,endpoint,timeout,port,env,service_account,expected", [
         pytest.param(
             [
-                turing.router.config.router_ensembler_config.EnvVar(
+                turing.router.config.common.EnvVar(
                     name="env_!@#$@%",
                     value="env_val")
             ],
@@ -472,12 +473,12 @@ def test_create_docker_router_ensembler_config_with_invalid_env(
             "500ms",
             5120,
             [
-                turing.router.config.router_ensembler_config.EnvVar(
+                turing.router.config.common.EnvVar(
                     name="env_name",
                     value="env_val")
             ],
             "secret-name-for-google-service-account",
-            turing.router.config.router_ensembler_config.InvalidEnvironmentVariableNameException
+            turing.router.config.common.InvalidEnvironmentVariableNameException
         )
     ])
 def test_set_docker_router_ensembler_config_with_invalid_env(
@@ -510,7 +511,7 @@ def test_set_docker_router_ensembler_config_with_invalid_env(
     "new_env,id,image,resource_request,endpoint,timeout,port,env,service_account,expected", [
         pytest.param(
             [
-                turing.router.config.router_ensembler_config.EnvVar(
+                turing.router.config.common.EnvVar(
                     name="env_name",
                     value="env_val")
             ],
@@ -526,7 +527,7 @@ def test_set_docker_router_ensembler_config_with_invalid_env(
             "500ms",
             5120,
             [
-                turing.router.config.router_ensembler_config.EnvVar(
+                turing.router.config.common.EnvVar(
                     name="env_not_the_right_name",
                     value="env_val")
             ],
